Moved component styles to new module file within components.
