+++ src/components/para_trials/timer.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useTimer } from "react-timer-hook";
-import styles from "../../styles/Paraflow.module.css";
+import styles from "./styles/Paratrials.module.css";
 
 interface TimerProps {
   timeInSec: number;
