+++ src/styles/Dashboard.module.css
@@ -1,5 +1,3 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter&family=Quicksand:wght@600&display=swap");
-
 /* CM DASHBOARD */
 .cmContainer {
   background-color: #f6f5ff;
