+++ src/styles/Paraflow.module.css
@@ -1,150 +1,8 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter&family=Quicksand:wght@600&display=swap");
-
-/* Styles for counter component */
-.counterContainer {
-  margin: 10px 0;
-  width: 100vw;
-  font-family: "Quicksand", sans-serif;
-}
-
-.counterButtonContainer {
-  display: grid;
-  grid-template-columns: 30% 40% 30%;
-  width: 80%;
-  height: 50px;
-  margin: 0 auto;
-  border-radius: 4px;
-  overflow: hidden;
-  font-weight: 600;
-  font-size: 40px;
-}
-
-.counterButton {
-  color: white;
-  font-size: 2rem;
-  width: 100%;
-  height: 100%;
-  border: none;
-  cursor: pointer;
-}
-
-.counterNumberDisplay {
-  color: black;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.counterButton-blue {
-  background-color: #1680A1;
-}
-.counterNumberDisplay-blue {
-  background-color: #1680A133;
-}
-.counterButton-green {
-  background-color: #167841;
-}
-.counterNumberDisplay-green {
-  background-color: #16784133;
-}
-.counterButton-yellow {
-  background-color: #FFAA44;
-}
-.counterNumberDisplay-yellow {
-  background-color: #FFAA4433;
-}
-
-.counterTitle {
-  font-size: 20px;
-  font-weight: 600;
-  color: #A2ACB3;
-  width: 100%;
-  text-align: center;
-}
-
-
-/* Styles for timer components */
-.timerContainer {
-  display: grid;
-  grid-template-columns: 1fr 2fr 1fr;
-  align-items: center;
-  justify-items: center;
-}
-
-.startStopButton {
-  border: none;
-  border-radius: 50%;
-  color: white;
-  background-color: #5347D7;
-  height: 50px;
-  width: 50px;
-  cursor: pointer;
-}
-
-.resetButton {
-  border: none;
+/* Styles for trials page (Could end up somewhere else) */
+.setTimerButton {
+  border-radius: 20px;
   background: none;
-  color: #20159E;
-  font-size: 50px;
-  cursor: pointer;
-}
-
-.timer {
-  font-size: 40px;
-  font-weight: 600;
-  display: flex;
-  color: #5347D7;
-}
-
-.timer span {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-}
-.timer span p {
-  font-size: 16px;
-  font-weight: normal;
-}
-
-.numPadContainer {
-  color: #5347D7;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.numPadContainer h2 {
-  color: black;
-}
-
-.numPad {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  grid-template-rows: repeat(4, minmax(0, 1fr));
-  row-gap: 5px;
-  column-gap: 5px;
-}
-
-.numPadButton {
-  color: inherit;
-  background: none;
-  border: none;
-  font-size: 32px;
-  font-weight: 600;
-  padding: 10px;
-  cursor: pointer;
-}
-
-.startButton {
-  border: none;
-  border-radius: 50%;
-  height: 100px;
-  width: 100px;
-  background-color: #5347D7;
-  color: white;
+  padding: 5px 10px;
+  border: 1px solid black;
   cursor: pointer;
 }
