+++ src/styles/globals.css
@@ -1,3 +1,5 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter&family=Quicksand:wght@600&display=swap");
+
 * {
   box-sizing: border-box;
   padding: 0;
